requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-markdown2==2.4.10
docling
python-dotenv==1.0.0
pydantic-settings==2.1.0
gunicorn==21.2.0
prometheus-client==0.19.0
python-jose[cryptography]==3.3.0
sentry-sdk[fastapi]==1.39.1
slowapi==0.1.8
